requirements.txt
pyTelegramBotAPI
pdfplumber
python-dotenv